menambah ui fitur di dashboard admin (Muhamad Rojali)

+++ backend/src/models/Order.js
@@ -6,7 +6,7 @@ const Order = sequelize.define(
   {
     id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
     user_id: { type: DataTypes.INTEGER, allowNull: false },
-    total_amount: { type: DataTypes.INTEGER, allowNull: false },
+    total_amount: { type: DataTypes.DECIMAL(12, 2), allowNull: false },
     status: {
       type: DataTypes.ENUM(
         "pending",
